Move summary to the first page of summary file.
Also removed summary from statistics file. Moving it to the top is not
possible without rewriting the file since statistics is written to it
during the experiments. This is to prevent data loss in case of a crash
or aborting. Moving the summery in front of the equation list hides it
some where in the middle of the statistics file. However, it is in its
own file, so there is no point in making it this complicated.